Tratamento de erro 404

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -19,6 +19,7 @@ import javax.validation.Valid;
 import java.net.URI;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/topicos")
@@ -52,25 +53,38 @@ public class TopicosController {
     }
 
     @GetMapping("/{id}")
-    public DetalhesDoTopicoDto detalhar(@PathVariable Long id){
-       Topico topico = topicoRepository.getOne(id);
-        return new DetalhesDoTopicoDto(topico);
+    public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id){
+       Optional<Topico> topico = topicoRepository.findById(id);
+       if(topico.isPresent()){
+           return ResponseEntity.ok( new DetalhesDoTopicoDto(topico.get()));
+       }
+
+       return ResponseEntity.notFound().build();
     }
 
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id,@RequestBody @Valid AtualizacaoTopicoForm form){
-            Topico topico = form.atualizar(id,topicoRepository);
+            Optional<Topico> opcional = topicoRepository.findById(id);
+            if(opcional.isPresent()){
+                Topico topico = form.atualizar(id,topicoRepository);
+                return ResponseEntity.ok(new TopicoDto(topico));
+        }
 
-            return ResponseEntity.ok(new TopicoDto(topico));
+
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity remover(@PathVariable Long id){
-        topicoRepository.deleteById(id);
+        Optional<Topico> opcional = topicoRepository.findById(id);
+        if(opcional.isPresent()){
+            topicoRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
 
-        return ResponseEntity.ok().build();
+        return ResponseEntity.notFound().build();
     }
 
 }
